CSS Styles for the Text Animation
This file holds the css code for the webkit, moz, and o animation that I called rainbow. It is what tells the browser to cycle through the colors of the text shadow.  It also holds the class "p" which is what the html code tag for the text itself is. Obviously, this animation can be called for any tag that will take text-shadow commands.